Revert "Ready for Netlify Deployment"
