--- netlify/functions/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "netlify-functions",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@google/genai": "^1.20.0",
-    "@types/aws-lambda": "^8.10.142",
-    "typescript": "~5.8.2"
-  }
-}
